fix(courses): address PR #30 review comments
- Move CreateCourseParams and UpdateCourseParams to types.ts (matches repo convention)
- Narrow UpdateCourseParams.default_view to union type matching Zod schema
- Add idempotentHint: true to update_course (PUT is idempotent)
- Update test to assert idempotentHint annotation
